Move type definitions to where they're actually used
